+++ src/config.ts
@@ -80,7 +80,7 @@ type ScoringConfig = {
     timeDecayExponent: number;
 };
 
-interface TagTootsConfig {
+export interface TagTootsConfig {
     maxToots: number;
     numTags: number;
     numTootsPerTag: number;