+++ src/scorer/feature/favourited_tags_scorer.ts
@@ -2,9 +2,8 @@
  * Score how many times the current user has favourited the toot's hashtags in the past.
  */
 import FeatureScorer from '../feature_scorer';
-import MastoApi from '../../api/api';
+import TagList from '../../api/tag_list';
 import Toot from '../../api/objects/toot';
-import { countTags } from '../../api/objects/tag';
 import { ScoreName, StringNumberDict } from '../../types';
 import { sumArray } from '../../helpers/collection_helpers';
 
@@ -17,7 +16,7 @@ export default class FavouritedTagsScorer extends FeatureScorer {
     };
 
     async prepareScoreData(): Promise<StringNumberDict> {
-        return countTags(await MastoApi.instance.getFavouritedToots());
+        return (await TagList.fromFavourites()).numTootsLookupDict();
     };
 
     async _score(toot: Toot) {