+++ src/scorer/feature/hashtag_participation_scorer.ts
@@ -2,8 +2,8 @@
  * Score toots containing hashtags the user posts about a lot
  */
 import FeatureScorer from '../feature_scorer';
+import TagList from '../../api/tag_list';
 import Toot from '../../api/objects/toot';
-import UserData from '../../api/user_data';
 import { ScoreName, StringNumberDict } from "../../types";
 import { sumArray } from '../../helpers/collection_helpers';
 
@@ -16,15 +16,7 @@ export default class HashtagParticipationScorer extends FeatureScorer {
     }
 
     async prepareScoreData(): Promise<StringNumberDict> {
-        const userTags = await UserData.getUserParticipatedTags();
-
-        return Object.values(userTags).reduce(
-            (acc, tag) => {
-                acc[tag.name] = tag.numToots || 0;
-                return acc;
-            },
-            {} as StringNumberDict
-        );
+        return (await TagList.fromParticipated()).numTootsLookupDict();
     };
 
     // Use the square root of the number of toots with the hashtag to prevent runaway scores